sysctl takes a 'size_t' variable for the size parameter
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-6-patches@19830 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/ucd-snmp/vmstat_bsdi4.c
@@ -155,9 +155,9 @@ var_extensible_vmstat(struct variable *vp,
 
     static struct vmmeter mem_old, mem_new;
     int             mem_mib[] = { CTL_VM, VM_CNT };
-    int             mem_size = sizeof(struct vmmeter);
+    size_t          mem_size = sizeof(struct vmmeter);
     int             cpu_mib[] = { CTL_KERN, KERN_CPUSTATS };
-    int             cpu_size = sizeof(struct cpustats);
+    size_t          cpu_size = sizeof(struct cpustats);
 
     static long     long_ret;
     static char     errmsg[300];
